+++ parser/src/main/scala/mdoc/parser/ParserSettings.scala
@@ -1,0 +1,5 @@
+package mdoc.parser
+
+trait ParserSettings {
+  val allowCodeFenceIndented: Boolean
+}
