+++ tests/unit/src/test/scala/tests/parser/MarkdownPartSuite.scala
@@ -1,43 +1,25 @@
-package tests.markdown
+package tests.parser
 
 import munit.{FunSuite, Location}
-import mdoc.internal.markdown.MarkdownFile
 
-import scala.meta.inputs.Input
-import mdoc.internal.io.ConsoleReporter
-import mdoc.internal.markdown.Text
-import mdoc.internal.markdown.MarkdownPart
-import mdoc.internal.markdown.CodeFence
+import mdoc.parser._
 
-import scala.meta.io.RelativePath
-import mdoc.internal.cli.InputFile
-
-import scala.meta.io.AbsolutePath
-import java.nio.file.Files
-import mdoc.internal.cli.Settings
-import mdoc.internal.markdown.MarkdownFile.Parser
-
-import scala.meta.internal.io.PathIO
-
-class MarkdownFileSuite extends BaseMarkdownSuite {
-  val reporter = new ConsoleReporter(System.out)
+class MarkdownPartSuite extends tests.BaseSuite {
+  private val parserSettings = new ParserSettings {
+    val allowCodeFenceIndented: Boolean = true
+  }
 
   def checkParse(name: String, original: String, expected: MarkdownPart*)(implicit
       loc: Location
   ): Unit = {
     test(name) {
-      reporter.reset()
-      val input = Input.VirtualFile(name, original)
-      val file = InputFile.fromRelativeFilename(name, Settings.default(PathIO.workingDirectory))
-      val obtained = MarkdownFile
-        .parse(input, file, reporter, baseSettings.copy(allowCodeFenceIndented = true))
-        .parts
-      require(!reporter.hasErrors)
+      val obtained = MarkdownPart.parse(original, parserSettings)
       val expectedParts = expected.toList
       assertNoDiff(
         pprint.tokenize(obtained).mkString,
         pprint.tokenize(expectedParts).mkString
       )
+      assertEquals(expected.length, obtained.length)
     }
   }
